chore: split log running tests
